changed grid geometry
In previous versions the grid structure was a square such that the
boundaries acted as walls, preventing repeaters (so called in the wiki
for Conway's Game of Life) from cascading across the grid.

Now the grid geometry is a torus, acting in the same exact way that the
arcade game Asteroids played. The top is connected to the bottom, and
the left is connected to the right.